refactor step definition
