+++ e2e/src/domains/category/domain/ports/category-ui.port.ts
@@ -7,13 +7,6 @@ export interface CategoryUiPort {
     parent?: string,
     expectError?: boolean
   ): Promise<string | void>;
-  createUniqueCategory(
-    name: string,
-    icon: string,
-    type: string,
-    parent?: string,
-    expectError?: boolean
-  ): Promise<string | void>;
   isCategoryCreated(name: string): Promise<boolean>;
   isCategoryChildOf(childName: string, parentName: string): Promise<boolean>;
   createCategoryWithDuplicateName(
