+++ e2e/src/domains/category/presentation/steps/navigation.steps.ts
@@ -1,0 +1,11 @@
+import { Given } from "@cucumber/cucumber";
+
+import { CustomWorld } from "../../../../support/world";
+
+Given(
+  "the user is on the Category Management page",
+  async function (this: CustomWorld) {
+    await this.getCategoryUseCase().navigateToCategoryPage();
+    await this.getCategoryUseCase().assertOnCategoryPage();
+  }
+);
